working

+++ src/vector_lib.h
@@ -6,7 +6,7 @@
 /*   By: awillems <awillems@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/17 15:33:10 by awillems          #+#    #+#             */
-/*   Updated: 2022/05/30 15:45:10 by awillems         ###   ########.fr       */
+/*   Updated: 2022/05/30 15:48:05 by awillems         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ t_vec	vec_init(size_t size);
 void	vec_delete(t_vec *vec);
 void	vec_destroy(t_vec *vec);
 
-void	*vec_get(t_vec *vec, size_t index);
+void	*vec_get(t_vec *vec, int index);
 
 t_vec	*vec_resize(t_vec *vec);
 t_vec	*vec_resize_round(t_vec *vec, size_t new_len);
